+++ libc/utils/hdrgen/hdrgen/enumeration.py
