+++ libc/utils/hdrgen/hdrgen/header.py
