+++ libc/utils/hdrgen/hdrgen/macro.py
