+++ libc/utils/hdrgen/hdrgen/object.py
