+++ libc/utils/hdrgen/hdrgen/type.py
