+++ libc/utils/hdrgen/hdrgen/yaml_functions_sorted.py
